requirements.txt
absl-py==1.2.0
aiohttp==3.8.6
aiosignal==1.3.1
anvil-uplink==0.3.42
anyio==3.6.1
appdirs==1.4.4
apyori==1.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.6.2
blis==0.7.9
cachetools==5.2.0
catalogue==1.0.2
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
datasets==2.14.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.6
Django==4.1.3
easyocr==1.6.0
editdistance==0.6.0
efficientnet==1.0.0
entrypoints==0.4
essential-generators==1.0
executing==0.8.3
fastapi==0.85.0
filelock==3.12.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
flatbuffers==2.0.7
fonttools==4.29.1
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.2
gast==0.4.0
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
googletrans==3.1.0a0
grpcio==1.49.1
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.17.3
hyperframe==5.2.0
idna==2.10
imageio==2.21.2
imgaug==0.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.10.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lab==7.0
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mysql-connector-python==8.0.28
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
neattext==0.1.2
nest-asyncio==1.5.4
networkx==2.8.6
ninja==1.10.2.3
nltk==3.7
notebook==6.4.8
numba==0.56.0
numpy==1.23.2
oauthlib==3.2.1
opencv-contrib-python==4.5.5.62
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.0
pickleshare==0.7.5
Pillow==9.0.1
plac==1.1.3
platformdirs==3.6.0
plotly==5.14.1
pmaw==3.0.0
pooch==1.6.0
praw==7.6.1
prawcore==2.3.0
preshed==3.0.8
print==1.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.0
psaw==0.1.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pycparser==2.21
pydantic==1.10.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
PySimpleGUI==4.60.4
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.7.24
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.1
rfc3986==1.5.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
Shapely==1.8.4
simplejson==3.17.6
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
SoundFile==0.10.3.post1
soupsieve==2.4.1
sqlparse==0.4.3
srsly==1.0.6
stack-data==0.2.0
starlette==0.20.4
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.13.3
testpath==0.6.0
thinc==7.4.6
threadpoolctl==3.1.0
tifffile==2022.8.12
tokenizers==0.14.1
torch==1.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.34.0
txt2tags==3.7
typing_extensions==4.3.0
tzdata==2022.6
Unidecode==1.3.6
update-checker==0.18.0
urllib3==1.26.11
uvicorn==0.18.3
validators==0.20.0
virtualenv==20.23.1
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.3.6
widgetsnbextension==3.5.2
wrapt==1.14.1
ws4py==0.5.1
xxhash==3.4.1
yarl==1.9.2